Completed from my side

+++ Core/Inc/Flash_Page_f0.h
@@ -23,7 +23,7 @@
 
 uint32_t Flash_Write_Data (uint32_t StartPageAddress, uint8_t *Data, uint16_t numberofwords);
 
-void Flash_Read_Data (uint32_t StartPageAddress, uint32_t *RxBuf, uint16_t numberofwords);
+void Flash_Read_Data (uint32_t StartPageAddress, uint8_t *RxBuf, uint16_t numberofwords);
 
 void Convert_To_Str (uint32_t *Data, char *Buf);
