+++ Core/Src/Flash_Page_f0.c
@@ -124,7 +124,7 @@ uint32_t Flash_Write_Data (uint32_t StartPageAddress, uint8_t *Data, uint16_t nu
 }
 
 
-void Flash_Read_Data (uint32_t StartPageAddress, uint32_t *RxBuf, uint16_t numberofwords)
+void Flash_Read_Data (uint32_t StartPageAddress, uint8_t *RxBuf, uint16_t numberofwords)
 {
 	while (1)
 	{
